+++ packages/cubejs-server-core/src/core/server.ts
@@ -107,6 +107,7 @@ export class CubejsServerCore {
 
   protected readonly orchestratorStorage: OrchestratorStorage = new OrchestratorStorage();
 
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
   protected repositoryFactory: ((context: RequestContext) => SchemaFileRepository) | (() => FileRepository);
 
   protected contextToDbType: DbTypeAsyncFn;
@@ -141,6 +142,7 @@ export class CubejsServerCore {
 
   protected apiGatewayInstance: ApiGateway | null = null;
 
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
   public readonly event: (name: string, props?: object) => Promise<void>;
 
   public projectFingerprint: string | null = null;
@@ -506,6 +508,7 @@ export class CubejsServerCore {
           ),
           externalDialectClass: this.options.externalDialectFactory && this.options.externalDialectFactory(context),
           schemaVersion: currentSchemaVersion,
+          contextToRoles: this.options.contextToRoles,
           preAggregationsSchema: await this.preAggregationsSchema(context),
           context,
           allowJsDuplicatePropsInSchema: this.options.allowJsDuplicatePropsInSchema,
@@ -667,6 +670,7 @@ export class CubejsServerCore {
       options.dbType || this.options.dbType,
       {
         schemaVersion: options.schemaVersion || this.options.schemaVersion,
+        contextToRoles: this.options.contextToRoles,
         devServer: this.options.devServer,
         logger: this.logger,
         externalDbType: options.externalDbType,